% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usar_tema_oscuro.R
\name{use_nighttime_theme}
\alias{use_nighttime_theme}
\title{Usar el tema oscuro.}
\usage{
use_nighttime_theme()
}
\description{
Cambiar rápidamente al tema oscuro, por defecto uso "R Espacial ES Dark".
}
